universal7580-common: update gpsd and gps.universal7580.so from T580XXS5CTD1.
As the Samsung Galaxy Tab A 10.1" 2016 has the same Broadcom BCM4752 GPS
chipset as in these devices, we can use the GPS blobs from its stock firmware.

Modifications made to gpsd to get it to work:
 * SSLv3_client_method has been replaced with SSLv23_method:
$ sed -i "s/SSLv3_client_method/SSLv23_method\x00\x00\x00\x00\x00\x00/" gpsd
 * The android.hidl.base@1.0 dependancy has been removed using patchelf:
$ patchelf --remove-needed android.hidl.base@1.0.so gpsd

gps.universal7580.so was taken directly from /vendor/lib/hw/gps.default.so
as-is from that firmware.